refac: moving functions from header files
To isolate the implementation code from the header files, I moved the
functions to new ".c" files in the src folder.
