+++ src/dimensionsProduct.h
@@ -1,12 +1,5 @@
 #ifndef DIMENSIONSPRODUCT_H
 #define DIMENSIONSPRODUCT_H
-
-int calculateDimensionsProduct(const int *arrayOfDimensions, const int subscriptPosition){
-	int dimensionsProduct = 1;
-	for(int i = 0; i < subscriptPosition; i++){
-		dimensionsProduct *= arrayOfDimensions[i];
-	}
-	return dimensionsProduct;
-}
+int calculateDimensionsProduct(const int *arrayOfDimensions, const int subscriptPosition);
 
 #endif
